+++ 2018/Sabella/main.c
@@ -12,6 +12,7 @@
 
 #include "formats.h"
 #include "analyzer.h"
+#include "libs/uthash/uthash.h"
 
 #define TRUE 1
 #define FALSE 0
@@ -25,6 +26,7 @@
 
 
 // ----- ----- STRUCTS ----- ----- //
+// Structure used to encapsulate threads' arguments
 struct _ThreadArgs {
   pcap_t** captureHandler;
   char* targetDev;
@@ -37,43 +39,58 @@ struct _ThreadArgs {
 };
 typedef struct _ThreadArgs ThreadArgs;
 
-struct _HostIP {
-  unsigned char hostIP[4];
+// Structure used to map network's host IP
+struct _HostMapEntry {
+  unsigned char ipAddress[4];
 
   unsigned long ID;
   UT_hash_handle hh; /* makes this structure hashable */
 };
-typedef struct _HostIP HostIP;
+typedef struct _HostMapEntry HostMapEntry;
 
 
 // ----- ----- GLOBALS ----- ----- //
 extern int errno;
 char gErrbuf[PCAP_ERRBUF_SIZE]; // pcap error buffer
 
+// Pcap's handlers and device used to sniff
+char* gTargetDev = NULL;
 pcap_t* gPoisongHandler;
 pcap_t* gGatekeeperHandler;
 
-char* gTargetDev = NULL;
+// Network basic informations
 char gGatewayIP[20]  = "192.168.1.1";
 unsigned char gGatewayMAC[6];
 const unsigned char gSpoofedMAC[6] = {0x3c, 0x5a, 0xb4, 0x88, 0x88, 0x88};
-u_char* gGratuitousARPSpoof;
 
-HostIP* gHostsMap = NULL; // Declaring hashtable
+uint8_t* gArpSpoofRequest; // The ARP request used to poison caches
+
+// Hash table used to map network's hosts
+HostMapEntry* gHostsMap = NULL; // Declaring hashtable
 pthread_mutex_t gHostMapLock = PTHREAD_MUTEX_INITIALIZER;
 
-int gTerminate    = FALSE; // Terminates the execution if something went wrong
-pthread_mutex_t m = PTHREAD_MUTEX_INITIALIZER;
+// Termination condition
+int gTerminate = FALSE;
+pthread_mutex_t gTerminationLock = PTHREAD_MUTEX_INITIALIZER;
 
 
 // ----- ----- FUNCTION DEFINITIONS ----- ----- //
-void  printUsage();
-void  breakLoops(int signum);
-char* policyParser(char* filename);
+// Prints the usage
+void  PrintUsage();
+// Setup several global variables and initializes almost all 'gArpSpoofRequest' fields
+void setup();
+// Toggles 'gTerminate' value to TRUE on SIGINT interruption
+void  SigIntHandler(int signum);
+// Parses the 'policy.txt' file
+char* PolicyParser(char* filename);
+// Sends a spoofed arp request (IP: gateway IP / MAC: spoofed MAC) to each host discovered in the network
 void  Spoof();
+// Initializes and handle a pcaploop
 void* pcapLoop(void *vargp);
-void  Poisoner(u_char *user, const struct pcap_pkthdr *h, const u_char *bytes);
-void  Gatekeeper(u_char *user, const struct pcap_pkthdr *h, const u_char *bytes);
+// Updates HostMap hash table using arp's src protocol address
+void  HostMapping(uint8_t *user, const struct pcap_pkthdr *h, const uint8_t *bytes);
+// Capture analyzes and filter traffic
+void  Gatekeeper(uint8_t *user, const struct pcap_pkthdr *h, const uint8_t *bytes);
 
 
 // ----- ----- ARGUMENT PARSING ----- ----- //
@@ -99,7 +116,7 @@ int parseUArgs(int argc, char const *argv[]) {
         break;
       }
       case 'h': {
-        printUsage();
+        PrintUsage();
         exit(0);
       }
       default:
@@ -121,35 +138,11 @@ int parseUArgs(int argc, char const *argv[]) {
 
 
 // ----- ----- MAIN ----- ----- //
-void setup() {
-  ARP_REQUEST = htons(ARP_REQUEST);
-  ARP_REPLY   = htons(ARP_REPLY);
-  IP_PROTO    = htons(IP_PROTO);
-
-  // Preparing Gratuoitous packet
-  unsigned char spoofIP[4];
-  stringToIP(gGatewayIP, spoofIP);
-
-  gGratuitousARPSpoof = (u_char*) malloc(ARP_SIZE);
-  EthHeader* eth = (EthHeader*) gGratuitousARPSpoof;
-  eth->type = htons(0x0806);
-  memcpy(eth->dstAddr, ETH_BROADCAST, 6);
-  memcpy(eth->srcAddr, gSpoofedMAC, 6);
-
-  ArpFormat* arp = (ArpFormat*) (gGratuitousARPSpoof + 14);
-  arp->operation = ARP_REQUEST;
-  arp->hwType = htons(0x0001); arp->hwLength = (6);
-  arp->prType = htons(0x0800); arp->prLenght = (4);
-
-  memcpy(arp->srcPrAddr, spoofIP, 4);
-  memcpy(arp->srcHwAddr, gSpoofedMAC, 6);
-  memcpy(arp->dstHwAddr, ETH_BROADCAST, 6);
-}
-
 int main(int argc, char const *argv[]) {
+  // ----- Initialization ----- //
   // Parsing policy file
   if(parseUArgs(argc, argv) == -1) return -1;
-  char* policies = policyParser("policy.txt");
+  char* policies = PolicyParser("policy.txt");
 
   setup();
 
@@ -167,7 +160,10 @@ int main(int argc, char const *argv[]) {
     return FAIL;
   }
 
-  // Poisoner loop
+  signal(SIGINT, SigIntHandler); // CRTL-C termination
+
+  // ----- Host Mapping ----- //
+  // Preparing HostMapping filter expression
   char* poisonerFilter = malloc(200 * sizeof(char));
   strcpy(poisonerFilter, "ether src host not ");
   char aus[20]; stringFyMAC(aus, gSpoofedMAC);
@@ -175,6 +171,7 @@ int main(int argc, char const *argv[]) {
   strcat(poisonerFilter, " && arp && arp src host not 0.0.0.0 && arp src host not ");
   strcat(poisonerFilter, gGatewayIP);
 
+  // Preparing HostMapping loop initializer arguments
   ThreadArgs* poisonArgs = malloc(sizeof(ThreadArgs));
   poisonArgs -> captureHandler = &gPoisongHandler;
   poisonArgs -> targetDev = gTargetDev;
@@ -183,16 +180,18 @@ int main(int argc, char const *argv[]) {
   poisonArgs -> packetBufferTm = 1;
   poisonArgs -> filterExp = poisonerFilter;
   poisonArgs -> devMask = devMask;
-  poisonArgs -> callback = Poisoner;
+  poisonArgs -> callback = HostMapping;
 
+  // Starting host mapping
   pthread_t poisonerTid;
   pthread_create(&poisonerTid, NULL, pcapLoop, (void*) poisonArgs);
 
-  // Gatekeeper loop
+  // ----- Gatekeeping ----- //
+  // Preparing Gatekeeper filter expression
   char spoofedMac[18];
   stringFyMAC(spoofedMac, gSpoofedMAC);
   int policiesLength = (policies == NULL) ? 0 : strlen(policies);
-  int filterLength = strlen(spoofedMac) + policiesLength + 100;
+  int filterLength = strlen(spoofedMac) + policiesLength + 150;
   char* gatekeeperFilter = malloc(filterLength * sizeof(char));
 
   strcat(gatekeeperFilter, "ether src not ");
@@ -206,6 +205,7 @@ int main(int argc, char const *argv[]) {
     free(policies);
   }
 
+  // Preparing Gatekeeper loop initializer arguments
   ThreadArgs* gatekeeperArgs = malloc(sizeof(ThreadArgs));
   gatekeeperArgs -> captureHandler = &gGatekeeperHandler;
   gatekeeperArgs -> targetDev = gTargetDev;
@@ -216,36 +216,87 @@ int main(int argc, char const *argv[]) {
   gatekeeperArgs -> devMask = devMask;
   gatekeeperArgs -> callback = Gatekeeper;
 
+  // Starting the gatekeeper
   pthread_t gatekeeperTid;
   pthread_create(&gatekeeperTid, NULL, pcapLoop, (void*) gatekeeperArgs);
 
-  signal(SIGINT, breakLoops); // CRTL-C termination
-
-  // If one of the threads stops exit
+  // ----- Poisoning cycle ----- //
   while (TRUE) {
-    pthread_mutex_lock(&m);
+    // Termination condition
+    pthread_mutex_lock(&gTerminationLock);
     int term = gTerminate;
-    pthread_mutex_unlock(&m);
+    pthread_mutex_unlock(&gTerminationLock);
     if(term) break;
 
+    // Spoofing then sleeping
     Spoof();
     usleep(500 * 1000);
   }
 
+  // ----- Closing ----- //
+  // Time to close, need stop threads
   if(gPoisongHandler != NULL) pcap_breakloop(gPoisongHandler);
   pthread_join(poisonerTid, NULL);
 
   if(gGatekeeperHandler != NULL) pcap_breakloop(gGatekeeperHandler);
   pthread_join(gatekeeperTid, NULL);
 
+  cleanAnalyzer(); // Cleaning analyzer
+
+  // Cleaning hosts hash map
+  if(gHostsMap != NULL) {
+    HostMapEntry *currentEntry, *tmp;
+
+    HASH_ITER(hh, gHostsMap, currentEntry, tmp) {
+      HASH_DEL(gHostsMap, currentEntry);
+      free(currentEntry);
+    }
+  }
+
   printf("\nGlad to stop, sir.\n");
-  cleanAnalyzer();
   return 0;
 }
 
 
+// ----- ----- ENVIRONMENT INITIALIZATION ----- ----- //
+void setup() {
+  // This variables have to be converted to network format
+  ARP_REQUEST = htons(ARP_REQUEST);
+  ARP_REPLY   = htons(ARP_REPLY);
+  IP_PROTO    = htons(IP_PROTO);
+
+  // Preparing spoofed request packet (see below for more info)
+  unsigned char spoofIP[4];
+  stringToIP(gGatewayIP, spoofIP);
+
+  gArpSpoofRequest = (uint8_t*) malloc(ARP_SIZE);
+  EthHeader* eth = (EthHeader*) gArpSpoofRequest;
+  eth->type = htons(0x0806);
+  memcpy(eth->dstAddr, ETH_BROADCAST, 6);
+  memcpy(eth->srcAddr, gSpoofedMAC, 6);
+
+  ArpMessage* arp = (ArpMessage*) (gArpSpoofRequest + 14);
+  arp->operation = ARP_REQUEST;
+  arp->hwType = htons(0x0001); arp->hwLength = (6);
+  arp->prType = htons(0x0800); arp->prLenght = (4);
+
+  memcpy(arp->srcPrAddr, spoofIP, 4);
+  memcpy(arp->srcHwAddr, gSpoofedMAC, 6);
+  memcpy(arp->dstHwAddr, ETH_BROADCAST, 6);
+}
+/*
+ * * * * ARP SPOOF REQUEST PACKET FORMAT * *
+ * OP:           request                    *
+ * dst hardware: ETH_BROADCAST              *
+ * src hardware: MAC used for spoofing      *
+ * dst protocol: host mapping dependent     *
+ * src protocol: target IP address          *
+ * * * * * * * * * * * * * * * * * * * * * *
+*/
+
+
 // ----- ----- MISCELLANEOUS ----- ----- //
-void printUsage() {
+void PrintUsage() {
   printf("\
 Usage: backfire [option]                                                       \n\
 Option:                                                                        \n\
@@ -256,13 +307,13 @@ Option:                                                                        \
   printf("\n");
 }
 
-void breakLoops(int signum) {
-  pthread_mutex_lock(&m);
+void SigIntHandler(int signum) {
+  pthread_mutex_lock(&gTerminationLock);
   gTerminate = TRUE;
-  pthread_mutex_unlock(&m);
+  pthread_mutex_unlock(&gTerminationLock);
 }
 
-char* policyParser(char* filename) {
+char* PolicyParser(char* filename) {
   FILE* fp;
   char* line = NULL;
   char* policy = NULL;
@@ -272,6 +323,7 @@ char* policyParser(char* filename) {
   fp = fopen(filename, "r");
   if (fp == NULL) return NULL;
 
+  // Reading file line by line
   while ((cread = getline(&line, &len, fp)) != -1) {
     line[strlen(line)-1] = '\0';
     if(strcmp(line, "") == 0) continue;
@@ -283,11 +335,13 @@ char* policyParser(char* filename) {
       strcat(policy, ")");
     }
     else {
+      // First line
       policy = malloc(cread * sizeof(char));
       strcpy(policy, line);
     }
   }
 
+  // Cleaning and returning
   fclose(fp);
   if (line) free(line);
   return policy;
@@ -295,10 +349,10 @@ char* policyParser(char* filename) {
 
 
 // ----- ----- LOOP INITIALIZER ----- ----- //
-void closeThread(ThreadArgs* args, pcap_t* handler, struct bpf_program* filter) {
-  pthread_mutex_lock(&m);
+void cleanThreadEnv(ThreadArgs* args, pcap_t* handler, struct bpf_program* filter) {
+  pthread_mutex_lock(&gTerminationLock);
   gTerminate = TRUE;
-  pthread_mutex_unlock(&m);
+  pthread_mutex_unlock(&gTerminationLock);
 
   if(handler != NULL) {
     pcap_close(handler);
@@ -315,70 +369,73 @@ void *pcapLoop(void *vargp) {
   // Creating and opening handler
   if((handler = pcap_open_live(args->targetDev, args->snaplen, args->promisc, args->packetBufferTm, gErrbuf)) == NULL) {
     fprintf(stderr, "Cannot create or open packet capture handle for %s\n", args->targetDev);
-    closeThread(args, handler, &filter);
+    cleanThreadEnv(args, handler, &filter);
     pthread_exit(NULL);
   }
+
+  // Setting up global handler
   pcap_t** ghandler = args->captureHandler;
   *ghandler = handler;
-
-  pcap_set_rfmon(handler, TRUE);
 
   // Setting up filters
   if (pcap_compile(handler, &filter, args->filterExp, TRUE, args->devMask) == -1) {
     fprintf(stderr, "Something wrong while parsing filter \"%s\": %s\n", args->filterExp, pcap_geterr(handler));
-    closeThread(args, handler, &filter);
+    cleanThreadEnv(args, handler, &filter);
     pthread_exit(NULL);
   }
 	if (pcap_setfilter(handler, &filter) == -1) {
     fprintf(stderr, "Something wrong while installing filter \"%s\": %s\n", args->filterExp, pcap_geterr(handler));
-    closeThread(args, handler, &filter);
+    cleanThreadEnv(args, handler, &filter);
     pthread_exit(NULL);
 	}
 
-  pthread_mutex_lock(&m);
+  // Checking if something went wrong in the other thread
+  pthread_mutex_lock(&gTerminationLock);
   int term = gTerminate;
-  pthread_mutex_unlock(&m);
+  pthread_mutex_unlock(&gTerminationLock);
   if(term == TRUE) {
-    closeThread(args, handler, &filter);
+    cleanThreadEnv(args, handler, &filter);
     pthread_exit(NULL);
   }
 
-  // Setting up callback
+  // Setting up callback and loopig
   int loopReturn = pcap_loop(handler, -1, args->callback, (unsigned char *) handler);
   if(loopReturn == -1) {
     fprintf(stderr, "Error while looping");
   }
 
-  // Cleaning
-  closeThread(args, handler, &filter);
+  // Cleaning and returning
+  cleanThreadEnv(args, handler, &filter);
   pthread_exit(NULL);
 }
 
 
-// ----- ----- POISONING ----- ----- //
+// ----- ----- POISONING && HOST MAPPING ----- ----- //
 void Spoof() {
-  ArpFormat* arp = (ArpFormat*) (gGratuitousARPSpoof + 14);
+  ArpMessage* arp = (ArpMessage*) (gArpSpoofRequest + 14);
 
+  // Looping throw all hash map entries
   pthread_mutex_lock(&gHostMapLock);
-  for(HostIP* host=gHostsMap; host!=NULL; host=(HostIP*) host->hh.next) {
-    memcpy(arp->dstPrAddr, host->hostIP, 4);
-    pcap_sendpacket(gPoisongHandler, gGratuitousARPSpoof, ARP_SIZE);
+  for(HostMapEntry* host=gHostsMap; host!=NULL; host=(HostMapEntry*) host->hh.next) {
+    memcpy(arp->dstPrAddr, host->ipAddress, 4);
+    pcap_sendpacket(gPoisongHandler, gArpSpoofRequest, ARP_SIZE);
   }
   pthread_mutex_unlock(&gHostMapLock);
 }
 
-void Poisoner(u_char *user, const struct pcap_pkthdr *h, const u_char *bytes) {
-  ArpFormat* arpMsg = (ArpFormat*) (bytes + 14); // 14 Byte Ethernet header
+void HostMapping(uint8_t *user, const struct pcap_pkthdr *h, const uint8_t *bytes) {
+  ArpMessage* arpMsg = (ArpMessage*) (bytes + 14); // 14 Byte Ethernet header
 
-  HostIP* host;
+  HostMapEntry* host;
   unsigned long hostID = hashString(arpMsg->srcPrAddr);
 
+  // Checking if hash table needs to be update
   pthread_mutex_lock(&gHostMapLock);
   HASH_FIND_INT(gHostsMap, &hostID, host);
   if(host == NULL) {
-    host = (HostIP*) malloc(sizeof(HostIP));
+    host = (HostMapEntry*) malloc(sizeof(HostMapEntry));
     host->ID = hostID;
-    memcpy(host->hostIP, arpMsg->srcPrAddr, 4);
+    memcpy(host->ipAddress, arpMsg->srcPrAddr, 4);
     HASH_ADD_INT(gHostsMap, ID, host);
   }
   pthread_mutex_unlock(&gHostMapLock);
@@ -386,7 +443,7 @@ void Poisoner(u_char *user, const struct pcap_pkthdr *h, const u_char *bytes) {
 
 
 // ----- ----- GATEKEEPING ----- ----- //
-void Gatekeeper(u_char *user, const struct pcap_pkthdr *h, const u_char *bytes) {
+void Gatekeeper(uint8_t *user, const struct pcap_pkthdr *h, const uint8_t *bytes) {
   Analyze(h->caplen, bytes);
 
   // Rebuilding Ethernet frame
